% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barwithucl.R
\name{barwithucl}
\alias{barwithucl}
\title{Display individual value with reference value of a variable.}
\usage{
barwithucl(x, iobs, ucl, plotname = "", ylabelname = "")
}
\arguments{
\item{x}{A vector witht the variable of interest.}

\item{iobs}{An integer with the observation index.}

\item{ucl}{A number with the horizontal reference value.}

\item{plotname}{Optional string with the plot title. Set to "" by default.}

\item{ylabelname}{Optional string with the y-axis label. Set to "" by default.}
}
\value{
Ggplot object with geom_col and geom_hline layers
}
\description{
Returns a ggplot object with the individual value of a variable as a geom_col
with an horizontal line reference.
}
\examples{
barwithucl(c(1:40), 20, max(c(1:40)))
barwithucl(c(1:40), 20, max(c(1:40)), plotname = "Example", ylabel = "x value")
}
