% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custombar.R
\name{custombar}
\alias{custombar}
\title{Customized bar plot}
\usage{
custombar(X, iobs, plotname = "", xlabelname = "Variables")
}
\arguments{
\item{X}{A matrix with the vector representing each observation stored in rows.}

\item{iobs}{An integer with the observation index.}

\item{plotname}{Optional string with the plot title. Set to  "" by default.}

\item{xlabelname}{Optional string with the x-axis label. Set to "Variables" by default.}
}
\value{
custbarob ggplot object with the geom_col layer and other customization layers.
}
\description{
\code{custombar} returns a ggplot object with the values of a vector with a customized geom_col layer.
The y-axis scale is built accounting the maximum absolute value in X. Therefore, the size of the columns is also
relative to the magnitude of the extreme values in X.
}
\examples{

}
