% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distplot.R
\name{distplot}
\alias{distplot}
\title{Distance plot}
\usage{
distplot(
  X,
  pcamodel,
  obstag = matrix(0, nrow(X), 1),
  plottitle = "Distance plot\\n"
)
}
\arguments{
\item{X}{A matrix with the data to be displayed.}

\item{pcamodel}{A list wiht the PCA model elements.}

\item{obstag}{Optional array with an integer for each observation used as a group tag.}

\item{plottitle}{Optional string with the plot title. Set to \code{"Distance plot"} by default.}
}
\value{
distplotobj ggplot object with the generated distance plot.
}
\description{
\code{distplot} returns a ggplot object with the distance plot for the observations of a matrix, expressing their coordinates
in terms of the Squared Prediction Error and the Hotelling's T^2 according to the provided pca model. It also includes the option
of distinguishing beteween observations and a customized title.
}
