% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscplot.R
\name{dscplot}
\alias{dscplot}
\title{dscplot}
\usage{
dscplot(
  X,
  pcamodel,
  obstag = matrix(0, nrow(X), 1),
  pcx = 1,
  pcy = 2,
  alpha = 0.05,
  nrow = 1,
  ncol = 2,
  legpos = "bottom"
)
}
\arguments{
\item{X}{Matrix with the data to be displayed.}

\item{pcamodel}{List wiht the PCA model elements.}

\item{obstag}{Optional column vector of integers indicating the group of each
observation (\code{0} or \code{1}). Default value set to \code{matrix(0, nrow(X), 1)}.}

\item{pcx}{Optional integer with the number of the PC in the horizontal axis. Set to \code{1} by default.}

\item{pcy}{Optional integer with the number of the PC in the vertical axis. Set to \code{2} by default.}

\item{alpha}{Optional number between 0 and 1 expressing the type I risk assumed in the compuatation of the confidence ellipse,
set to \code{0.05} (5 \%) by default.}

\item{nrow}{Optional number of rows the plot layout. Set to \code{1} by default.}

\item{ncol}{Optional number of columns the plot layout. Set to \code{2} by default.}

\item{legpos}{Optional string with the position of the legend. Set to \code{"bottom"}
by default.}
}
\value{
ggplot object with the generated score plot.
}
\description{
Returns the distance plot and the score plot providing a dataset and a pca model. Observations can
be identified by the obstag input argument.
}
