% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obscontribpanel.R
\name{obscontribpanel}
\alias{obscontribpanel}
\title{Information about T^2 and SPE statistics of an observation.}
\usage{
obscontribpanel(obsid, pcaref, pcax)
}
\arguments{
\item{obsid}{Integer with the index of the observation of interest.}

\item{pcaref}{A list with the PCA model according to which the distance and contributions are expressed.}

\item{pcax}{A list with the elements of the PCA model that will be displayed: SPE, T^2_A and their constributions (E and T2matrix)}
}
\value{
contrplots ggplot object with the generated bar plots.
}
\description{
\code{obscontribpanel} returns a ggplot object with four subplots showing the information of an observation regarding
its T^2 and SPE statistics, i.e.: information about the distance of the observation in the latent space and its distance to the model
hyperplane. On one hand, it presents a bar plot indicating the value of each statistic for the observation, referenced by their
Upper Control Limita. On the other hand, it also presents the contributions of that observation both for the T^2 and the SPE.
}
