% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcame.R
\name{pcame}
\alias{pcame}
\title{pcame}
\usage{
pcame(X, pcaref)
}
\arguments{
\item{X}{Matrix with observations that will be projected onto the PCA model.}

\item{pcaref}{A list with the elemements of a PCA model: \code{m} (mean), \code{s} (standard deviation), \code{prepro} (preprocessing:
\code{"none"}, \code{"cent"} or \code{"autosc"}), \code{P} (loading matrix), \code{lambda} (vector with variances of each PC).}
}
\value{
list with elements containing information about X in the PCA model: \code{Xpreprocessed} (matrix \code{X} preprocessed),
\code{Tscores} (score matrix with the projection of \code{X} on each one of the A PCs), \code{E} (error matrix with the par of \code{X}
not explained by the PCA model), \code{SPE} (vector with the SPE for each observation of \code{X}), \code{T2} (vector with the T^_A for
each observation of \code{X}), \code{T2matrix} (matrix with the contributions of each PC to the T^2_A for each observation of \code{X})
and \code{Xrec} (matrix with the reconstructed part of \code{X}, i.e. the part of \code{X} explained by the PCA model).
}
\description{
Projection of X onto a PCA model.
}
\details{
\code{pcame} performs the projection of the data in X onto the PCA model stored as a list of parameters. It returns the projection
of the observations in X, along with the SPE, Hotelling's T^2_A, contribution elements and the reconstruction of X obtained by the
PCA model.
}
