% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speinfo.R
\name{speinfo}
\alias{speinfo}
\title{Information about Squared Prediction Error (SPE) for an observation.}
\usage{
speinfo(SPE, E, limspe, iobs = iobs)
}
\arguments{
\item{SPE}{A vector with values of the SPE statistic.}

\item{E}{A matrix with the contributions of each variable (columns) for each observation (rows) to the SPE. It is the error term obtained
from the unexplained part of X by the PCA model.}

\item{iobs}{Integer with the index of the observation of interest.}

\item{limht2}{A number with the value of the Upper Control Limit for the SPE (at a certain confidence level (1-alpha)*100 \%).}
}
\value{
speplots ggplot object with the generated bar plots.
}
\description{
\code{speinfo} returns a ggplot object with two subplots showing the information of an observation regarding its SPE  statistic, i.e.:
information about the distance of the observation to the the PCA model hyperplane. On one hand, it presents a bar plot indicating
the value of the statistic for the observation, referenced by the Upper Control Limit. On the other hand, it also presents
the contribution that each component had for the SPE value of the selected observation (i.e. its error vector).
}
