% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distplotsimple.R
\name{distplotsimple}
\alias{distplotsimple}
\title{Distance plot}
\usage{
distplotsimple(
  T2,
  SPE,
  lim.t2,
  lim.spe,
  ncomp,
  obstag = matrix(0, length(T2), 1),
  alpha = 0.05,
  plottitle = "Distance plot\\n"
)
}
\arguments{
\item{T2}{Vector with the Hotelling's T^2 values for each observation.}

\item{SPE}{Vector with the SPE values for each observation.}

\item{lim.t2}{Value of the Upper Control Limit for the T^2 statistic.}

\item{lim.spe}{Value of the Upper Control Limit for the SPE.}

\item{obstag}{Optional array with an integer for each observation used as a group tag. Set to \code{matrix(0, length(T2), 1)}
by default.}

\item{alpha}{Optional number between 0 and 1 expressing the type I risk assumed in the compuatation of the confidence ellipse,
set to 0.05 (5 \%) by default.}

\item{plottitle}{Optional string with the plot title, \code{"Distance plot"} by default.}
}
\value{
distplotobj ggplot object with the generated distance plot.
}
\description{
Returns a ggplot object expressing the observations according to their Hotelling's T^2 (T^2) and
Squared Prediction Error (SPE) coordinates. It also includes the Upper Control Limit references of each statistic, along with
the option of distinguishing beteween observations and a customized title.
}
