% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcame.R
\name{pcame}
\alias{pcame}
\title{Projection of X onto a PCA model.}
\usage{
pcame(X, pcamodel)
}
\arguments{
\item{X}{A vector or matrix with observations that will be projected onto the PCA model.}

\item{pcamodel}{A list with the elemements of a PCA model: m (mean), s (standard deviation), prepro (preprocessing),
P (loading matrix), lambda (vector with variances of each PC). The preprocessing element is a character with possible values "none",
if any preprocessing should be performed on X, "cent", if a mean-centering should be performed on X, or "autosc", it a mean-centering
and unitary variance scaling (autoscaling) should be performed on X.}
}
\value{
pcaout list with fields containing information about X in the PCA model: Xpreprocessed (matrix X preprocessed),
Tscores (score matrix with the projection of X on each one of the A PCs), E (error matrix with the par of X not explained by the PCA model),
SPE (vector with the SPE for each observation of X), T2 (vector with the T^_A for each observation of X), T2matrix (matrix with the contributions of each PC
to the T^2_A for each observation of X) and Xrec (matrix with the reconstructed part of X, i.e. the part of X explained by the PCA model).
}
\description{
\code{pcame} performs the projection of the data in X onto the PCA model stored as a list of parameters. It returns the projection
of the observations in X, along with the SPE, Hotelling's T^2_A, contribution elements and the reconstruction of X obtained by the
PCA model.
}
