% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutsimple.R
\name{scoutsimple}
\alias{scoutsimple}
\title{Shift of an array with a single step.}
\usage{
scoutsimple(X, pcaref, T2.target = NA, SPE.target = NA)
}
\arguments{
\item{X}{A vector or matrix with observations that will be shifted as rows.}

\item{pcaref}{A list with the elemements of a PCA model: m (mean), s (standard deviation), prepro (preprocessing),
P (loading matrix), lambda (vector with variances of each PC). The preprocessing element is a character with possible values "none",
if any preprocessing should be performed on X, "cent", if a mean-centering should be performed on X, or "autosc", it a mean-centering
and unitary variance scaling (autoscaling) should be performed on X.}

\item{T2.y}{A number indicating the target value for the T^2_A after the shift. Set to NA by default.}

\item{SPE.y}{A number indicating the target value for the SPE after the shift. Set to NA by default.}
}
\value{
scoutsimple list with elements with information about the shifted data. The matrix X, contains the new data, the SPE and T2 contain
the statistic values of each one of the new generated outliers or observations. The elements step.spe and step.t2 make reference to the step
at which each observation of the shifted dada is located. Finally, the element tag, has different values for each observation in X according
to each step (if tag.mode == "series"), otherwise is a vector of ones with as many elements as rows in X.
}
\description{
\code{scoutsimple} returns a new list with the information about the shifts performed to the observations in X.
}
