% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speinfo.R
\name{speinfo}
\alias{speinfo}
\title{Information about the Squared Prediction Error (SPE) of an observation. Two subplots show the information of an observation regarding its
SPE statistic, i.e.: a bar plot indicating the value of the statistic for the observation, and a bar plot with the contribution
that each component had for the SPE value}
\usage{
speinfo(SPE, E, limspe, iobs = NA)
}
\arguments{
\item{SPE}{Vector with values of the SPE statistic.}

\item{E}{Matrix with the contributions of each variable (columns) for each observation (rows) to the SPE. It is the error term obtained
from the unexplained part of X by the PCA model.}

\item{limspe}{Upper Control Limit for the SPE, at a certain confidence level (1-alpha)*100 \%.}

\item{iobs}{Integer with the index of the observation of interest. Default value set to \code{NA}.}
}
\value{
ggplot object with the generated bar plots.
}
\description{
Information about the Squared Prediction Error (SPE) of an observation. Two subplots show the information of an observation regarding its
SPE statistic, i.e.: a bar plot indicating the value of the statistic for the observation, and a bar plot with the contribution
that each component had for the SPE value
}
