% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X}
\alias{X}
\title{Demo dataset with 50 observations and 5 normally distributed variables with
two Principal Components explaining the 80\% of the total variance.}
\format{
A matrix data frame with 50 rows and 5 normally distributed variables.
}
\usage{
X
}
\description{
It is a small data set to use as a demo for the SCOUTer package.
}
\keyword{datasets}
