% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xshift.R
\name{xshift}
\alias{xshift}
\title{Shift of an observation.}
\usage{
xshift(X, P, a, b)
}
\arguments{
\item{X}{Vector or matrix with observations that will be shifted}

\item{P}{Loading matrix of the PCA model according to which the shfit will be performed.}

\item{a}{A number or vector tuning the amount of the shift for each observation from X in the direction of the x's projection.}

\item{b}{A number tuning the amount of the shift for each observation from X in the direction of the x's residual.}
}
\value{
Xout matrix with shifted observation as rows in the same order as in the input matrix (X).
}
\description{
\code{xshift} returns a new array, shifted accordingly to the factors a and b which determine the direction of the shift. The performed
operation results as a combination of two main directions: the direction of maximum gradient for the SPE (weighted by the parameter b) and
the direction of the projection of the observation on the model (weighted by the parameter a).
}
